Move withName function to a bigger scope
