+++ processor/src/main/java/com/infinum/jsonapix/processor/specs/TypeAdapterListSpecBuilder.kt
@@ -3,7 +3,7 @@ package com.infinum.jsonapix.processor.specs
 import com.infinum.jsonapix.core.JsonApiModel
 import com.infinum.jsonapix.core.adapters.TypeAdapter
 import com.infinum.jsonapix.core.common.JsonApiConstants
-import com.infinum.jsonapix.core.common.JsonApiConstants.Prefix.withName
+import com.infinum.jsonapix.core.common.JsonApiConstants.withName
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
